duratioin改用int

+++ app/models.py
@@ -119,7 +119,7 @@ class Video(Base):
     song_id: Mapped[int] = mapped_column(Integer, ForeignKey('song.id'), nullable=False)
     copyright: Mapped[int] = mapped_column(SmallInteger, nullable=True)
     thumbnail: Mapped[str] = mapped_column(Text, nullable=True)
-    duration: Mapped[int] = mapped_column(SmallInteger, nullable=True)
+    duration: Mapped[int] = mapped_column(Integer, nullable=True)
     page: Mapped[int] = mapped_column(SmallInteger, nullable=True)
 
     uploader: Mapped["Uploader"] = relationship("Uploader", back_populates="videos")
